+++ src/Drupal/Component/DrupalTwigComponent.php
@@ -1,0 +1,18 @@
+<?php
+
+/*
+ * This file is part of Mannequin.
+ *
+ * (c) 2017 Last Call Media, Rob Bayliss <rob@lastcallmedia.com>
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
+namespace LastCall\Mannequin\Drupal\Component;
+
+use LastCall\Mannequin\Twig\Component\TwigComponent;
+
+class DrupalTwigComponent extends TwigComponent
+{
+}